refactor(console): move api creation v4 to directory v4
