+++ gravitee-apim-console-webui/src/management/api/creation/v4/steps/step-1/api-creation-v4-step-1.component.ts
@@ -24,7 +24,7 @@ import { ApiDetails } from '../../models/api-details';
 @Component({
   selector: 'api-creation-v4-step-1',
   template: require('./api-creation-v4-step-1.component.html'),
-  styles: [require('./api-creation-v4-step-1.component.scss'), require('../api-creation-v4.component.scss')],
+  styles: [require('./api-creation-v4-step-1.component.scss'), require('../../api-creation-v4.component.scss')],
 })
 export class ApiCreationV4Step1Component implements OnInit {
   public form: FormGroup;
