+++ gravitee-apim-console-webui/src/management/api/creation/v4/steps/step-2/api-creation-v4-step-2.component.ts
@@ -32,7 +32,7 @@ type EntrypointVM = {
 @Component({
   selector: 'api-creation-v4-step-2',
   template: require('./api-creation-v4-step-2.component.html'),
-  styles: [require('./api-creation-v4-step-2.component.scss'), require('../api-creation-v4.component.scss')],
+  styles: [require('./api-creation-v4-step-2.component.scss'), require('../../api-creation-v4.component.scss')],
 })
 export class ApiCreationV4Step2Component implements OnInit, OnDestroy {
   private unsubscribe$: Subject<void> = new Subject<void>();
